feat: implement SW05's fibe large universe scheme

+++ fibe/utils/check_attributes_array.go
@@ -1,6 +1,6 @@
 package utils
 
-func CheckAttributesArray(attributes []int, universe int) bool {
+func CheckAttributesArray(attributes []int64, universe int64) bool {
 	if attributes == nil || len(attributes) == 0 {
 		return false
 	}
